Consolidating photoatomic blocks in their own directories
