+++ cmake/unit_testing.cmake
@@ -80,10 +80,10 @@ add_subdirectory( src/ACEtk/block/ThermalScatteringDiscreteCosines/test )
 add_subdirectory( src/ACEtk/block/ThermalScatteringDiscreteCosinesWithProbability/test )
 add_subdirectory( src/ACEtk/block/ThermalScatteringElasticAngularDistributionBlock/test )
 add_subdirectory( src/ACEtk/block/ThermalScatteringInelasticAngularDistributionBlock/test )
-add_subdirectory( src/ACEtk/block/PhotoatomicPrincipalCrossSectionBlock/test )
-add_subdirectory( src/ACEtk/block/IncoherentScatteringFunctionBlock/test )
-add_subdirectory( src/ACEtk/block/CoherentFormFactorBlock/test )
-add_subdirectory( src/ACEtk/block/PhotoatomicFluorescenceDataBlock/test )
+add_subdirectory( src/ACEtk/photoatomic/PrincipalCrossSectionBlock/test )
+add_subdirectory( src/ACEtk/photoatomic/IncoherentScatteringFunctionBlock/test )
+add_subdirectory( src/ACEtk/photoatomic/CoherentFormFactorBlock/test )
+add_subdirectory( src/ACEtk/photoatomic/FluorescenceDataBlock/test )
 add_subdirectory( src/ACEtk/block/PhotoatomicHeatingNumbersBlock/test )
 add_subdirectory( src/ACEtk/block/PhotoatomicComptonProfile/test )
 add_subdirectory( src/ACEtk/block/PhotoatomicComptonProfileBlock/test )
