+++ python/src/photoatomic/CoherentFormFactorBlock.python.cpp
@@ -3,7 +3,7 @@
 #include <pybind11/stl.h>
 
 // local includes
-#include "ACEtk/block/CoherentFormFactorBlock.hpp"
+#include "ACEtk/photoatomic/CoherentFormFactorBlock.hpp"
 #include "views.hpp"
 #include "definitions.hpp"
