+++ python/test/CMakeLists.txt
@@ -62,10 +62,10 @@ add_test( NAME ACEtk.python.block.ThermalScatteringDiscreteCosines COMMAND ${PYT
 add_test( NAME ACEtk.python.block.ThermalScatteringDiscreteCosinesWithProbability COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_ThermalScatteringDiscreteCosinesWithProbability.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
 add_test( NAME ACEtk.python.block.ThermalScatteringElasticAngularDistributionBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_ThermalScatteringElasticAngularDistributionBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
 add_test( NAME ACEtk.python.block.ThermalScatteringInelasticAngularDistributionBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_ThermalScatteringInelasticAngularDistributionBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
-add_test( NAME ACEtk.python.block.PhotoatomicPrincipalCrossSectionBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotoatomicPrincipalCrossSectionBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
-add_test( NAME ACEtk.python.block.IncoherentScatteringFunctionBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_IncoherentScatteringFunctionBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
-add_test( NAME ACEtk.python.block.CoherentFormFactorBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_CoherentFormFactorBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
-add_test( NAME ACEtk.python.block.PhotoatomicFluorescenceDataBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotoatomicFluorescenceDataBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
+add_test( NAME ACEtk.python.photoatomic.PrincipalCrossSectionBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/photoatomic/Test_ACEtk_PhotoatomicPrincipalCrossSectionBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
+add_test( NAME ACEtk.python.photoatomic.IncoherentScatteringFunctionBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/photoatomic/Test_ACEtk_IncoherentScatteringFunctionBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
+add_test( NAME ACEtk.python.photoatomic.CoherentFormFactorBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/photoatomic/Test_ACEtk_CoherentFormFactorBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
+add_test( NAME ACEtk.python.photoatomic.FluorescenceDataBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/photoatomic/Test_ACEtk_PhotoatomicFluorescenceDataBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
 add_test( NAME ACEtk.python.block.PhotoatomicHeatingNumbersBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotoatomicHeatingNumbersBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
 add_test( NAME ACEtk.python.block.PhotoatomicComptonProfile COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotoatomicComptonProfile.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
 add_test( NAME ACEtk.python.block.PhotoatomicComptonProfileBlock COMMAND ${PYTHON_EXECUTABLE} -m unittest -v test/block/Test_ACEtk_PhotoatomicComptonProfileBlock.py WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/python )
